validation + exception handling done

+++ EmployeeMicroService/src/main/java/com/example/emp/service/EmployeeService.java
@@ -1,12 +1,12 @@
 package com.example.emp.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.example.emp.entity.Employee;
+import com.example.emp.exceptions.EmployeeNotFoundException;
 import com.example.emp.repository.EmployeeRepository;
 
 @Service
@@ -16,20 +16,32 @@ public class EmployeeService {
 	EmployeeRepository repo;
 
 	public List<Employee> getAllEmps() {
+
 		return repo.findAll();
 	}
 
 	public void addemp(Employee emp) {
-		 repo.save(emp);
+		repo.save(emp);
 	}
 
 	public void deleteEmp(int id) {
-		 repo.deleteById(id);
+		
+		Employee emp = repo.findById(id).orElse(null);
+		if (emp == null) 
+		{
+			throw new EmployeeNotFoundException("employeeId " + id + " not avilable to delete");
+		}
+		repo.deleteById(id);
 	}
 
-	public Optional<Employee> getEmpByid(int id) {
-		
-		return repo.findById(id);
+	public Employee getEmpByid(int id) {
+
+		Employee emp = repo.findById(id).orElse(null);
+		if (emp == null) {
+			throw new EmployeeNotFoundException("employeeId " + id + " not avilable to get");
+		}
+		return emp;
+
 	}
-	
+
 }
